+++ threedi_modelchecker/checks/raster.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from math import isclose
 from pathlib import Path
 from typing import Optional, Set, Type
 
@@ -207,9 +208,11 @@ class RasterGridSizeCheck(BaseRasterCheck):
             if not raster.is_valid_geotiff:
                 return True
             # the x pixel size is used here,but it is equal to the y pixel size
-            return ((self.grid_space / raster.pixel_size[0]) % 2 == 0) and (
-                self.grid_space >= (2 * raster.pixel_size[0])
-            )
+            return (
+                isclose(
+                    a=((self.grid_space / raster.pixel_size[0]) % 2), b=0, rel_tol=1e-09
+                )
+            ) and (self.grid_space >= (2 * raster.pixel_size[0]))
 
     def description(self):
         return "v2_global_settings.grid_space is not a positive even multiple of the raster cell size."