+++ threedi_modelchecker/__init__.py
@@ -1,5 +1,5 @@
 from .model_checks import *  # NOQA
 
 # fmt: off
-__version__ = '2.2.3.dev0'
+__version__ = '2.2.3'
 # fmt: on